+++ apptentive-core/src/main/java/apptentive/com/android/core/ApptentiveException.kt
@@ -1,0 +1,6 @@
+package apptentive.com.android.core
+
+import apptentive.com.android.util.InternalUseOnly
+
+@InternalUseOnly
+open class ApptentiveException(message: String, cause: Throwable? = null) : Exception(message, cause)