+++ apptentive-navigate-to-link/src/main/java/apptentive/com/android/feedback/link/view/NavigateTolinkActivity.kt
@@ -1,16 +1,34 @@
 package apptentive.com.android.feedback.link.view
 
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
+import android.webkit.ValueCallback
+import android.webkit.WebChromeClient
 import android.webkit.WebView
+import androidx.activity.result.ActivityResultLauncher
+import androidx.activity.result.contract.ActivityResultContracts
 import apptentive.com.android.feedback.link.R
 import apptentive.com.android.ui.hideSoftKeyboard
+import apptentive.com.android.util.Log
+import apptentive.com.android.util.LogTags
 import com.google.android.material.appbar.MaterialToolbar
 
 internal class NavigateTolinkActivity : BaseNavigateToLinkActivity() {
     private lateinit var webView: WebView
+    private var uploadMessage: ValueCallback<Array<Uri>>? = null
+    private lateinit var fileChooserLauncher: ActivityResultLauncher<Intent>
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.apptentive_activity_navigate_to_link)
+
+        // Register the ActivityResultLauncher
+        fileChooserLauncher = registerForActivityResult(ActivityResultContracts.StartActivityForResult()) { result ->
+            if (uploadMessage == null) return@registerForActivityResult
+            uploadMessage?.onReceiveValue(WebChromeClient.FileChooserParams.parseResult(result.resultCode, result.data))
+            uploadMessage = null
+        }
 
         supportActionBar?.hide()
 
@@ -21,7 +39,42 @@ internal class NavigateTolinkActivity : BaseNavigateToLinkActivity() {
         }
 
         webView = findViewById<WebView>(R.id.apptentive_webview_navigate_to_link)
-        webView.settings.javaScriptEnabled = true
+        val settings = webView.settings
+        settings.javaScriptEnabled = true
+        settings.domStorageEnabled = true
+        settings.mediaPlaybackRequiresUserGesture = false
+        settings.javaScriptCanOpenWindowsAutomatically = true
+        webView.webChromeClient = object : WebChromeClient() {
+            override fun onShowFileChooser(
+                webView: WebView?,
+                filePathCallback: ValueCallback<Array<Uri>>?,
+                fileChooserParams: FileChooserParams?
+            ): Boolean {
+                if (uploadMessage != null) {
+                    uploadMessage?.onReceiveValue(null)
+                    uploadMessage = null
+                }
+
+                Log.i(LogTags.SURVEY, "Detected file upload using alchemer survey")
+
+                uploadMessage = filePathCallback
+                val intent = fileChooserParams?.createIntent()
+                try {
+                    fileChooserLauncher.launch(intent)
+                } catch (e: Exception) {
+                    Log.e(LogTags.SURVEY, "Error launching file chooser", e)
+                    uploadMessage = null
+                    return false
+                }
+                return true
+            }
+
+            override fun onReceivedTitle(view: WebView?, title: String?) {
+                super.onReceivedTitle(view, title)
+                title?.let { topAppBar.title = it }
+            }
+        }
+
         val url = intent.getStringExtra("linkUrl")
         if (savedInstanceState != null) {
             webView.restoreState(savedInstanceState)
@@ -32,5 +85,16 @@ internal class NavigateTolinkActivity : BaseNavigateToLinkActivity() {
     override fun onSaveInstanceState(outState: Bundle) {
         super.onSaveInstanceState(outState)
         webView.saveState(outState)
+    }
+
+    override fun onDestroy() {
+        super.onDestroy()
+        webView.apply {
+            clearHistory()
+            clearCache(true)
+            onPause()
+            removeAllViews()
+            destroy()
+        }
     }
 }